Two buffer fixes that was overwriting the buffer
The first bug, fixed by Jim Ley would occur when the buffer reached 1024 and
the parser as parsing a size byte.

The second fix is in regards to how the timestamp is created.
The time stamp would loop 3 times (hr/sec/min) and then continue.
However, if the buffer was near 512 bytes or 1024 bytes, it would
not write to disk instead would clobber the other buffer.

Both fixed.